@@ -14,53 +14,58 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.2.4",
-    "@angular/common": "^4.2.4",
-    "@angular/compiler": "^4.2.4",
-    "@angular/core": "^4.2.4",
-    "@angular/forms": "^4.2.4",
-    "@angular/http": "^4.2.4",
-    "@angular/platform-browser": "^4.2.4",
-    "@angular/platform-browser-dynamic": "^4.2.4",
-    "@angular/router": "^4.2.4",
-    "angular2-froala-wysiwyg": "^2.6.0",
+    "@angular/animations": "^7.2.10",
+    "@angular/common": "^7.2.10",
+    "@angular/compiler": "^7.2.10",
+    "@angular/core": "^7.2.10",
+    "@angular/forms": "^7.2.10",
+    "@angular/http": "^7.2.10",
+    "@angular/platform-browser": "^7.2.10",
+    "@angular/platform-browser-dynamic": "^7.2.10",
+    "@angular/router": "^7.2.10",
+    "angular-froala-wysiwyg": "^2.9.3",
     "animate.css": "^3.5.2",
     "bootstrap": "^3.3.7",
-    "core-js": "^2.4.1",
+    "core-js": "^2.5.4",
     "echarts": "^3.8.5",
     "font-awesome": "^4.7.0",
+    "fontface-source-sans-pro": "^2.0.10-alpha1",
     "jquery": "^3.2.1",
     "ng2-file-upload": "^1.3.0",
     "ng2-select": "^2.0.0",
     "ngx-echarts": "^2.0.0",
-    "ngx-modal": "0.0.29",
+    "ngx-modal": "^0.0.29",
     "ngx-pagination": "^3.0.3",
+    "node-sass": "^4.12.0",
     "normalize.css": "^7.0.0",
     "pell": "^1.0.4",
     "roboto-fontface": "^0.8.0",
-    "rxjs": "^5.4.2",
+    "rxjs": "^6.4.0",
+    "rxjs-compat": "^6.0.0-rc.0",
     "sweetalert2": "^7.6.3",
-    "zone.js": "^0.8.14"
+    "tslib": "^1.9.0",
+    "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.3.1",
-    "@angular/compiler-cli": "^4.2.4",
-    "@angular/language-service": "^4.2.4",
-    "@types/jasmine": "~2.5.53",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "~3.1.1",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
+    "@angular-devkit/build-angular": "~0.13.0",
+    "@angular/cli": "^7.3.6",
+    "@angular/compiler-cli": "^7.2.10",
+    "@angular/language-service": "^7.2.10",
+    "@types/jasmine": "~2.8.8",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.9.4",
+    "codelyzer": "~4.5.0",
+    "jasmine-core": "~2.99.1",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~4.0.0",
+    "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~3.2.0",
-    "tslint": "~5.3.2",
-    "typescript": "~2.3.3"
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.11.0",
+    "typescript": "~3.2.2"
   }
 }
