@@ -6,9 +6,9 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./profile.component.scss']
 })
 export class ProfileComponent implements OnInit {
-  avatarImgSrc: string = 'assets/images/avatar.png';
-  userName: string = 'Folisise Chosielie';
-  userPost: string = 'Musician, Player';
+  avatarImgSrc: string = 'assets/images/avatar.jpg';
+  userName: string = 'Bknds';
+  userPost: string = 'FrontEnd';
   
   constructor() { }
 
