@@ -1,6 +1,6 @@
-import { Component, Input, Output, EventEmitter, OnInit } from '@angular/core';
-import { menuService } from '../../services/menu.service';
-import { GlobalService } from '../../services/global.service';
+import {Component, Input, Output, EventEmitter, OnInit, HostListener, AfterViewInit} from '@angular/core';
+import {menuService} from '../../services/menu.service';
+import {GlobalService} from '../../services/global.service';
 
 @Component({
   selector: 'sidebar',
@@ -8,14 +8,15 @@ import { GlobalService } from '../../services/global.service';
   styleUrls: ['./sidebar.component.scss'],
   providers: [menuService]
 })
-export class SidebarComponent implements OnInit {
+export class SidebarComponent implements OnInit, AfterViewInit {
 
 
   public menuInfo: Array<any> = [];
   public sidebarToggle = true;
 
   constructor(private _menuService: menuService,
-    public _globalService: GlobalService) { }
+              public _globalService: GlobalService) {
+  }
 
   ngOnInit() {
     this.menuInfo = this._menuService.putSidebarJson();
@@ -57,4 +58,16 @@ export class SidebarComponent implements OnInit {
     }
   }
 
+
+  // detect window size and automatically hide the left side menu
+  @HostListener('window:resize', ['$event'])
+  onResize(event) {
+    this.sidebarToggle = window.innerWidth >= 970;
+    console.log(window.innerWidth);
+  }
+
+  ngAfterViewInit(): void {
+    this.sidebarToggle = window.innerWidth >= 970;
+    console.log(window.innerWidth);
+  }
 }
